+++ src/handler/index.ts
@@ -1,0 +1,2 @@
+export { scheduled } from './scheduled';
+export type { Handler } from './type';
